Fix threefold repetition.
Background: zobrist hashing is used for two different purposes: detecting repeated board states for threefold repetition, and looking up previous searches in the move tree. For the purposes of the latter, we wanted to distinguish board states by whose turn it is and which bugs the pillbug can throw, so there are extra bits xored in for those.

Threefold repetition was broken in two different ways that cancelled each other out for both test cases.
* Extra bits encoding which bug was moved last were not applied for bug placement. This undercounted repetitions where the first board state was the placement of pieces.
* Counting to three was broken, as it was counting to two including the final board state.

Now we count to 2 not including the final board state, and keep the pillbug movement bits out of the zobrist history. It is now lazily computed for anything else calling the public zobrist_hash().

Fixes #5.